Create  U2.A5.FOR_1
An elementary school needs a system to calculate grades, but depending on the grade, the subjects vary and that generates a more complex job to be able to calculate the average of each student correctly.

The director wants the teachers to be able to enter the name of the student they are going to evaluate, then how many subjects they are going to grade, then, enter each grade. If the result is greater than 6, it should show a message that he/she passed, if it is less, it should show a message that he/she needs to be retaken and if the grade is equal to 10, it should show a message that he/she is an outstanding student.

At the end, the program should show a report with the name of the student, his/her final average and the message that will depend on his/her grade. For delivery, attach the code in a text file (. txt)

------------------------------------------------------------------------