Create U2.A6.FOR_2
A shoemaker needs to record the data of its salespeople and the data of individual sales to calculate some quality standards to improve productivity and give bonuses, as well as "cut off heads."

The system should ask how many salespeople are to be recorded and for each salesperson it should ask how many sales they made and how much each was worth.

After finishing the sales record, it should show the total sales for that salesperson, the number of sales they made, and their highest sale. If the highest sale is less than $100 or the total is less than $1,500, it should display a message to go to Human Resources.

The final report should give a report of how many salespeople were registered in the system and how much the total global sales of the company was. If the global report is greater than $20,000 it should display a message that the goal was reached, otherwise it will display a message that a meeting should be called.

------------------------------------------------